Edit task 4-print_rev.c project 0x05

+++ 0x05-pointers_arrays_strings/4-print_rev.c
@@ -1,6 +1,6 @@
 #include "main.h"
 
-int _lenstr(char *s);
+unsigned long int _lenstr(char *s);
 
 /**
  * print_rev - print string s in a reversed order followed by new line
